requirements.txt
cloudpickle==2.0.0
cmake==3.22.0
cycler==0.11.0
fonttools==4.28.2
graphviz==0.19
gym==0.21.0
-e git+https://github.com/openai/retro.git@98fe0d328e1568836a46e5fce55a607f47a0c332#egg=gym_retro
kiwisolver==1.3.2
matplotlib==3.5.0
neat-python==0.92
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
pyglet==1.5.21
pyparsing==3.0.6
python-dateutil==2.8.2
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
